synchronously flush semantics when semantics is enabled
